feat: implement autocomplete
Created an input component which while typing suggests to users what
options of what they might want to write.

For this project, the starwars API is used. It auto-suggests character
name while user is searching for one.

Below the AutoComplete component is a list of Suggestions user's can
pick and click on to complete what they want to write.